README: Add no changing pin loc / origin symbol rule
To avoid situations where this git repo is updated causing schematic
nets to break due to symbol pin locations changing unexpectedly.  KiCad
does  not cache the symbols within the schematic itself (which is unlike
how footprints are often cached within a layout itself so that in order
to cause a change to a footprint that the newer version of a footprint
must be imported by the user in an explicit operation).

Signed-off-by: Andrew Bradford <andrew@bradfordembedded.com>